OBJ-70 Move IUpoadResponderStrategy to separate file
