+++ src/controllers/document.upload.controller.ts
@@ -12,7 +12,15 @@ import {
 import logger from "../utils/logger";
 import { MAX_FILE_SIZE_BYTES } from "../utils/properties";
 import { uploadFile, UploadFileCallbacks } from "./upload/http.request.file.uploader";
-import { createUploadResponderStrategy, IUploadResponderStrategy } from "./upload/upload.responder.strategy.factory";
+import { IUploadResponderStrategy } from "./upload/upload.responder.strategy";
+import { createUploadResponderStrategy } from "./upload/upload.responder.strategy.factory";
+
+/**
+ * Document upload controller
+ * Handles uploading of files
+ * Uses the strategy pattern to handle responses back to the browser differently for AJAX and non AJAX requests
+ * AJAX requests are sent from a javascript file included into the web page - upload.js, stored in the CDN
+ */
 
 /**
  * Handle GET request for document upload page
@@ -30,17 +38,17 @@ export const get = async (req: Request, res: Response) => {
 };
 
 /**
- * Handle POST request for document upload page
+ * Handle POST request for document upload page when a file is submitted
  * @param {Request} req the http request
  * @param {Response} res the http response
  * @param {NextFunction} next the next function in the middleware chain
  */
 export const postFile = async (req: Request, res: Response, next: NextFunction) => {
-  const isAjaxRequest: boolean = req.xhr;
-  logger.debugRequest(req, "Add document http request type is " + (isAjaxRequest ? "" : "NOT ") + "AJAX / XmlHttpRequest");
+  logger.debugRequest(req, "Add document http request type is " + (isAjaxRequest(req) ? "" : "NOT ") + "AJAX / XmlHttpRequest");
 
-  const uploadResponderStrategy: IUploadResponderStrategy = createUploadResponderStrategy(isAjaxRequest);
+  const uploadResponderStrategy: IUploadResponderStrategy = createUploadResponderStrategy(isAjaxRequest(req));
 
+  // TODO - try catch handlegeneric
   const attachments = getAttachments(req.session as Session);
 
   const uploadCallbacks: UploadFileCallbacks = {
@@ -48,15 +56,21 @@ export const postFile = async (req: Request, res: Response, next: NextFunction) 
     noFileDataReceivedCallback: getNoFileDataReceivedCallback(req, res, uploadResponderStrategy, attachments),
     uploadFinishedCallback: getUploadFinishedCallback(req, res, next, uploadResponderStrategy),
   };
-  const maxSizeBytes: number = parseInt(MAX_FILE_SIZE_BYTES, 10);
+  const maxFileSizeBytes: number = parseInt(MAX_FILE_SIZE_BYTES, 10);
 
-  return uploadFile(req, maxSizeBytes, uploadCallbacks);
+  return uploadFile(req, maxFileSizeBytes, uploadCallbacks);
 };
 
-export const postContinueButton = async (req: Request, res: Response, next: NextFunction) => {
-  const isAjaxRequest: boolean = req.xhr;
-  const uploadResponderStrategy: IUploadResponderStrategy = createUploadResponderStrategy(isAjaxRequest);
+/**
+ * Handle POST request for document upload page when continue button is pressed
+ * @param {Request} req the http request
+ * @param {Response} res the http response
+ * @param {NextFunction} next the next function in the middleware chain
+ */
+export const postContinueButton = async (req: Request, res: Response) => {
+  const uploadResponderStrategy: IUploadResponderStrategy = createUploadResponderStrategy(isAjaxRequest(req));
 
+  // TODO - try catch? handlegeneric
   const attachments = getAttachments(req.session as Session);
 
   if (attachments && attachments.length === 0) {
@@ -65,6 +79,14 @@ export const postContinueButton = async (req: Request, res: Response, next: Next
   res.redirect("TODO - PAGE AFTER UPLOAD");
 };
 
+/**
+ * Get uploadFile callback function for file size limit exceeded event
+ * @param {Request} req http request
+ * @param {Response} res http response
+ * @param {IUploadResponderStrategy} uploadResponderStrategy the strategy for responding to requests
+ * @param {any[]} attachments the list of attachments
+ * @returns {((filename: string, maxInBytes: number): void)} the callback function
+ */
 const getFileSizeLimitExceededCallback = (req: Request,
                                           res: Response,
                                           uploadResponderStrategy: IUploadResponderStrategy,
@@ -77,6 +99,14 @@ const getFileSizeLimitExceededCallback = (req: Request,
   };
 };
 
+/**
+ * Get uploadFile callback function for no file data received
+ * @param {Request} req http request
+ * @param {Response} res http response
+ * @param {IUploadResponderStrategy} uploadResponderStrategy the strategy for responding to requests
+ * @param {any[]} attachments the list of attachments
+ * @returns {((filename: string): void)} the callback function
+ */
 const getNoFileDataReceivedCallback = (req: Request,
                                        res: Response,
                                        uploadResponderStrategy: IUploadResponderStrategy,
@@ -86,6 +116,14 @@ const getNoFileDataReceivedCallback = (req: Request,
   };
 };
 
+/**
+ * Get uploadFile callback function for the upload finished event
+ * @param {Request} req http request
+ * @param {Response} res http response
+ * @param {NextFunction} next the next function in the middleware chain
+ * @param {IUploadResponderStrategy} uploadResponderStrategy the strategy for responding to requests
+ * @returns {((filename: string, fileData: Buffer, mimeType: string): void)} the callback function
+ */
 const getUploadFinishedCallback = (req: Request,
                                    res: Response,
                                    next: NextFunction,
@@ -117,17 +155,37 @@ const getUploadFinishedCallback = (req: Request,
   };
 };
 
-const displayError = async (req: Request,
+/**
+ * Displays an error message on the page
+ * @param {Request} req http request
+ * @param {Response} res http response
+ * @param {string} errorMessage the error message to display on the page
+ * @param {IUploadResponderStrategy} uploadResponderStrategy the strategy for responding to requests
+ * @param {any[]} attachments the list of attachments
+ */
+const displayError = async (req: Request, // TODO is this needed?
                             res: Response,
                             errorMessage: string,
                             uploadResponderStrategy: IUploadResponderStrategy,
-                            attachments: any) => {
+                            attachments: any[]) => {
   const documentUploadErrorData: GovUkErrorData =
     createGovUkErrorData(errorMessage, "#file-upload", true, "");
   return uploadResponderStrategy.handleGovUKError(res, documentUploadErrorData, attachments);
 };
 
-// Gets max file size in MB rounded down to nearest whole number
+/**
+ * Gets max file size in MB rounded down to nearest whole number
+ * @param {number} maxSizeInBytes the max size allowed in bytes
+ * @returns number max size in MB
+ */
 const getMaxFileSizeInMB = (maxSizeInBytes: number): number => {
   return Math.floor(maxSizeInBytes / (1024 * 1024));
+};
+
+/**
+ * Is this an Ajax request or not
+ * @param {boolean} req http request
+ */
+const isAjaxRequest = (req: Request): boolean => {
+  return req.xhr;
 };
