+++ src/controllers/upload/html.upload.responder.strategy.ts
@@ -2,7 +2,7 @@ import { NextFunction, Request, Response } from "express";
 import { GovUkErrorData } from "../../model/govuk.error.data";
 import { OBJECTIONS_DOCUMENT_UPLOAD } from "../../model/page.urls";
 import { Templates } from "../../model/template.paths";
-import { IUploadResponderStrategy } from "./upload.responder.strategy.factory";
+import { IUploadResponderStrategy } from "./upload.responder.strategy";
 
 export class HtmlUploadResponderStrategy implements IUploadResponderStrategy {
