+++ src/controllers/upload/upload.responder.strategy.factory.ts
@@ -1,16 +1,13 @@
-import { NextFunction, Request, Response } from "express";
-import { GovUkErrorData } from "../../model/govuk.error.data";
 import { AjaxUploadResponderStrategy } from "./ajax.upload.responder.strategy";
 import { HtmlUploadResponderStrategy } from "./html.upload.responder.strategy";
+import { IUploadResponderStrategy } from "./upload.responder.strategy";
 
-export interface IUploadResponderStrategy {
-  handleSuccess(req: Request, res: Response): void;
-  handleGenericError(res: Response, e: Error, next?: NextFunction): void;
-  handleGovUKError(res: Response, errorData: GovUkErrorData, attachments: any[]): void;
-}
-
-export const createUploadResponderStrategy = (isXhr: boolean): IUploadResponderStrategy => {
-  if (isXhr) {
+/**
+ * Factory method for creating an uploadResponderStrategy
+ * @param {boolean} isAjaxRequest do we need an ajax responder or not
+ */
+export const createUploadResponderStrategy = (isAjaxRequest: boolean): IUploadResponderStrategy => {
+  if (isAjaxRequest) {
     return new AjaxUploadResponderStrategy();
   } else {
     return new HtmlUploadResponderStrategy();
