+++ test/controller/upload/ajax.upload.responder.strategy.spec.unit.ts
@@ -3,7 +3,7 @@ jest.mock("../../../src/utils/logger");
 
 import { Request, Response } from "express";
 import { AjaxUploadResponderStrategy } from "../../../src/controllers/upload/ajax.upload.responder.strategy";
-import { IUploadResponderStrategy } from "../../../src/controllers/upload/upload.responder.strategy.factory";
+import { IUploadResponderStrategy } from "../../../src/controllers/upload/upload.responder.strategy";
 import { ErrorMessages } from "../../../src/model/error.messages";
 import { createGovUkErrorData, GovUkErrorData } from "../../../src/model/govuk.error.data";
 import { OBJECTIONS_ERROR } from "../../../src/model/page.urls";
