--- test/resources/files/text_large.txt
@@ -1,1 +1,0 @@
-Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Proin sed libero enim sed faucibus turpis in eu mi. Pretium aenean pharetra magna ac placerat. At consectetur lorem donec massa sapien faucibus et. Purus semper eget duis at tellus at urna condimentum mattis. Felis eget nunc lobortis mattis aliquam faucibus. Pellentesque id nibh tortor id aliquet. Mauris augue neque gravida in fermentum et sollicitudin ac. Amet risus nullam eget felis eget nunc. Ut aliquam purus sit amet luctus venenatis lectus magna. Faucibus in ornare quam viverra. Eu sem integer vitae justo eget magna fermentum. Feugiat in ante metus dictum at. Fames ac turpis egestas integer.
